Restructure redux boiler
