+++ src/Reducers/hasErrored.js
@@ -1,0 +1,8 @@
+export const hasErrored = (state = '', action) => {
+  switch (action.type) {
+    case 'HAS_ERRORED':
+      return action.msg
+    default:
+      return state
+  }
+}
