+++ src/Reducers/isLoading.js
@@ -1,0 +1,8 @@
+export const isLoading = (state = false, action) => {
+  switch (action.type) {
+    case 'IS_LOADING':
+      return action.isLoading
+    default:
+      return state
+  }
+}
